Initial commit.
Initial commit of a PHP implementation of the Aho-Corasick string matching
algorithm, packaged for distribution as a standalone library. The motivation
behind this work is the desire to replace FastStringSearch with a pure-PHP
implementation with comparable performance and without the licensing issues
which encumber FastStringSearch.

The Aho-Corasick string matching algorithm is an efficient algorithm for
locating occurrences of multiple search keywords in a body of text. The time
it takes to construct the finite state machine is proportional to the sum of
the lengths of the search keywords. Once constructed, the machine can locate
all occurences of all search keywords in a body of text in a single pass.

The library is made up of three classes:

* MultiStringMatcher: the base implementation.
* DeterministicMultiStringMatcher: a variant implementation which uses a
  deterministic finite automaton, as described in section 6 of the Aho-Corasick
  paper ("Eliminating failure transitions"). This variant computes all possible
  state transitions in advance, which then allows the automaton to make exactly
  one state transition per input character. The cost is having to do more work
  up-front.
* MultiStringReplacer: The MultiStringReplacer class extends
  MultiStringMatcher, adding search-and-replace capabilities with an interface
  similar to strtr() (when invoked with an array argument).

 [1]: http://cr.yp.to/bib/1975/aho.pdf

Change-Id: I583e1c2f619fa2b3992ac0909164d8a1fe8dc4d2